// src/api/parameters/getActionsIdMember.ts
export interface GetActionsIdMember {
  /** The API key to use */
  key: Record<string, any>;
  /** The API token to use */
  token: Record<string, any>;
  /** The ID of the Action */
  id: Record<string, any>;
  /** `all` or a comma-separated list of member fields */
  fields?: Record<string, any>;
}